Add Neumann and Robin BCs for XFEM interface
This PR defines a base class for XFEM interface kernels `XFEMInterfaceKernel` that derives from DiracKernel to provide boilerplate for integration points, weights and interface normals. Four classes derive from the `XFEMInterfaceKernel`: `XFEMInterfaceNeumannBC`, `XFEMInterfaceFunctionNeumannBC`, `XFEMInterfaceConvectiveHeatFluxBC` and `XFEMPressure`. Their names are self explanatory.

A geometric cut userobject need to be provided to the objects to tell the interface kernel where to perform the integration.

closes #17331